fix(parent): serialize fees_remaining as float not Decimal string
Pydantic's default Decimal serialization emits JSON strings ('90000.00')
but the FE Zod schema validates fees_remaining as z.number(). The mismatch
made the dashboard fetch fail validation client-side, surfacing as
'Connexion au serveur impossible' even though the BE returned 200.

Switched to float (precision loss negligible — we display thousands of XOF,
not micropayments) and explicit cast in service. Ship together with the
endpoint introduction so prod never sees the broken contract.

Discovered via /visual-check post-deploy of #121.

Refs #120

+++ app/schemas/parent_portal.py
@@ -53,7 +53,11 @@ class ParentDashboardChild(BaseModel):
     class_name: str
     general_average: float | None
     total_absences: int
-    fees_remaining: Decimal
+    # Float (not Decimal) so the JSON encoder emits a number, not a string.
+    # The FE Zod schema validates as z.number() — Pydantic's default
+    # Decimal→str serialization breaks the contract. Acceptable precision
+    # loss because we display thousands of XOF, not micropayments.
+    fees_remaining: float
 
 
 class ParentDashboardResponse(BaseModel):
